MCP_Client_Gemini/requirements.txt
google-genai>=0.5.0
requests>=2.28.0
mcp[cli]